Reference layer wip
