+++ PixiEditor/Views/UserControls/Layers/LayerGroupControl.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="PixiEditor.Views.UserControls.LayerGroupControl"
+﻿<UserControl x:Class="PixiEditor.Views.UserControls.Layers.LayerGroupControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
