+++ PixiEditor/Views/UserControls/Layers/LayerItem.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="PixiEditor.Views.LayerItem"
+﻿<UserControl x:Class="PixiEditor.Views.UserControls.Layers.LayerItem"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
