+++ PixiEditor/Views/UserControls/Layers/RawLayersViewer.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="PixiEditor.Views.UserControls.RawLayersViewer"
+﻿<UserControl x:Class="PixiEditor.Views.UserControls.Layers.RawLayersViewer"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
